comment-stock-api

+++ api/Controllers/StockController.cs
@@ -25,16 +25,22 @@ namespace api.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAllAsync()
+        public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stocks = await _stockRepo.GetAllAsync();
             var stocksDto = stocks.Select(stock => stock!.ToStockDto());
             return Ok(stocksDto);
         }
 
-        [HttpGet("{id}")]
-        public async Task<IActionResult> GetByIdAsync([FromRoute] int id)
+        [HttpGet("{id:int}")]
+        public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stock = await _stockRepo.GetByIdAsync(id);
 
             if (stock == null)
@@ -44,21 +50,27 @@ namespace api.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> CreateAsync([FromBody] CreateStockRequestDto createDto)
+        public async Task<IActionResult> Create([FromBody] CreateStockRequestDto createDto)
         {
-            var stockModel = createDto.ToStockFromCreateDto();
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var stockModel = createDto.ToStockFromCreate();
             await _stockRepo.CreateAsync(stockModel); 
 
             return CreatedAtAction(
-                nameof(GetByIdAsync),
+                nameof(GetById),
                 new { id = stockModel.Id },
                 stockModel.ToStockDto());
         }
 
         [HttpPut]
-        [Route("{id}")]
-        public async Task<IActionResult> UpdateAsync([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
+        [Route("{id:int}")]
+        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
 
             if (stockModel == null) 
@@ -68,9 +80,12 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
-        public async Task<IActionResult> DeleteAsync([FromRoute] int id)
+        [Route("{id:int}")]
+        public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+                
             var stockModel = await _stockRepo.DeleteAsync(id);
 
             if (stockModel == null)
